refactor pipeline into an impl package to reduce circular imports
